ReadOnlyLedgerOwnerBuilder
